Split main.rs into multiple files
